Refactored weight handling to grams and removed obsolete OrderBloc
This commit transitions the application's weight measurement system from floating-point kilograms to integer-based grams to ensure precision in calculations and storage. It also removes the unused `OrderBloc`, refactors rating logic, and improves UI form handling.

Key changes:
-   **Data Models & Precision:**
    -   **Catch & Offer Models:** Converted weight fields (`initialWeight`, `availableWeight`, `weight`, `previousWeight`) from `double` to `int` (grams).
    -   **Review Model:** Updated `ratingValue` to be stored as an `int`.
    -   **Converters:** Added `formatWeight` helper to display integer gram values as formatted Kg strings.

-   **Business Logic & Repositories:**
    -   **Price Calculation:** Updated `OfferRepository` and form logic to calculate `pricePerKg` using integer arithmetic: `(Total Price * 1000) / Weight(g)`.
    -   **Cleanup:** Deleted the unused `features/fisher/logic/order_bloc` directory (Bloc, Event, State).

-   **UI & Forms:**
    -   **Input Handling:** Refactored `OfferDetails`, `ProductDetails`, `CatchDetails`, and `CounterOfferDialog` to convert user input (Kg) to grams before submission and dynamically auto-calculate pricing fields.
    -   **Notifications Screen:** Implemented a `TabController` with badges for "Offers" and "Messages" counts.
    -   **Ratings:** Updated `AnimatedRatingStars` to handle integer ratings and fixed interaction logic.
    -   **Display Components:** Updated `ForSaleCard`, `OfferCard`, `SoldCard`, and `OrderCard` to utilize the new `formatWeight` utility.

-   **Maintenance:**
    -   **Seeder:** Updated `CatchSeeder` to generate test data with weights in grams.
    -   **Config:** Bumped Dart SDK environment constraint to `^3.10.0`.

+++ lib/features/user/presentation/widgets/review_card.dart
@@ -6,7 +6,7 @@ import 'package:siren_marketplace/core/widgets/error_handling_circle_avatar.dart
 class ReviewCard extends StatelessWidget {
   final String name;
   final String date;
-  final double rating;
+  final int rating;
   final String image;
   final String message;
 
@@ -45,9 +45,9 @@ class ReviewCard extends StatelessWidget {
                     children: [
                       // Star Icons
                       AnimatedRatingStars(
-                        initialRating: rating.toDouble(),
+                        initialRating: rating,
                         minRating: 1,
-                        maxRating: 5.0,
+                        maxRating: 5,
 
                         filledColor: AppColors.shellOrange,
                         emptyColor: AppColors.gray100,
